Added files via upload
This file allows users to drag in the canvas and it will record the difference between their pos every 16 milliseconds.  The sensitivity still needs to be implemented as well as passing the variable.   This file is pretty well documented so hopefully anyone can pick it up and try to fix these issues.